+++ test/metaevent.jl
@@ -31,4 +31,22 @@
             @test_throws errtype MIDI.readsysexevent(Int(input[1]), IOBuffer(input[2:length(input)]))
         end
     end
+
+    @testset "SequencerSpecificEvent" begin
+        midi = load("SequencerSpecific.mid")
+        sse = midi.tracks[1].events[1]
+        @test sse isa MIDI.SequencerSpecificEvent
+        sse.ssdata == [0x11, 0x21, 0x53, 0x1F]
+
+        sse = MIDI.SequencerSpecificEvent(0, 0x7f, [0x11, 0x21, 0x53, 0x1F])
+        @test sse.ssdata == [0x11, 0x21, 0x53, 0x1F]
+    end
+
+    @testset "check empty" begin
+        midi = load(testmidi())
+        track = midi.tracks[1]
+        @test !isempty(track)
+        empty!(track)
+        @test isempty(track)
+    end
 end